# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidGenius API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export type LoginRequestRole =
  (typeof LoginRequestRole)[keyof typeof LoginRequestRole];

export const LoginRequestRole = {
  admin: "admin",
  bidder: "bidder",
} as const;

export interface LoginRequest {
  email: string;
  password: string;
  role?: LoginRequestRole;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  bidder: "bidder",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  totalCredits: number;
  reservedCredits: number;
  availableCredits: number;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message?: string;
}

export type AuctionStatus = (typeof AuctionStatus)[keyof typeof AuctionStatus];

export const AuctionStatus = {
  upcoming: "upcoming",
  active: "active",
  ended: "ended",
} as const;

export interface Auction {
  id: number;
  title: string;
  description: string;
  imageUrl?: string | null;
  startTime: string;
  endTime: string;
  minBid: number;
  bidIncrement: number;
  currentBid: number;
  highestBidderId?: number | null;
  highestBidderName?: string | null;
  status: AuctionStatus;
  winnerId?: number | null;
  winnerName?: string | null;
  totalBids: number;
  createdAt: string;
}

export interface Bid {
  id: number;
  auctionId: number;
  userId: number;
  userName: string;
  amount: number;
  createdAt: string;
}

export type AuctionDetail = Auction & {
  bids: Bid[];
};

export interface CreateAuctionRequest {
  title: string;
  description: string;
  imageUrl?: string | null;
  startTime: string;
  endTime: string;
  /** @minimum 1 */
  minBid: number;
  /** @minimum 1 */
  bidIncrement: number;
}

export interface UpdateAuctionRequest {
  title?: string;
  description?: string;
  imageUrl?: string | null;
}

export interface PlaceBidRequest {
  /** @minimum 1 */
  amount: number;
}

export type BidWithAuctionAuctionStatus =
  (typeof BidWithAuctionAuctionStatus)[keyof typeof BidWithAuctionAuctionStatus];

export const BidWithAuctionAuctionStatus = {
  upcoming: "upcoming",
  active: "active",
  ended: "ended",
} as const;

export type BidWithAuction = Bid & {
  auctionTitle: string;
  auctionStatus: BidWithAuctionAuctionStatus;
  isHighestBid: boolean;
};

export type BidRecommendationStrategy =
  (typeof BidRecommendationStrategy)[keyof typeof BidRecommendationStrategy];

export const BidRecommendationStrategy = {
  safe: "safe",
  aggressive: "aggressive",
  moderate: "moderate",
} as const;

export interface BidRecommendation {
  suggestedBid: number;
  strategy: BidRecommendationStrategy;
  reasoning: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  confidence: number;
}

export type AssignCreditsRequestAction =
  (typeof AssignCreditsRequestAction)[keyof typeof AssignCreditsRequestAction];

export const AssignCreditsRequestAction = {
  add: "add",
  set: "set",
} as const;

export interface AssignCreditsRequest {
  /** @minimum 1 */
  amount: number;
  action?: AssignCreditsRequestAction;
}

export interface AdminStats {
  totalAuctions: number;
  activeAuctions: number;
  endedAuctions: number;
  upcomingAuctions: number;
  totalBidders: number;
  totalBids: number;
  recentActivity: Bid[];
}

export interface GenerateDescriptionRequest {
  title: string;
  keywords?: string;
}

export interface GenerateDescriptionResponse {
  description: string;
}

export type GetAuctionsParams = {
  status?: GetAuctionsStatus;
};

export type GetAuctionsStatus =
  (typeof GetAuctionsStatus)[keyof typeof GetAuctionsStatus];

export const GetAuctionsStatus = {
  upcoming: "upcoming",
  active: "active",
  ended: "ended",
} as const;