# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BidGenius API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auctions:
    get:
      operationId: getAuctions
      tags: [auctions]
      summary: Get all auctions
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [upcoming, active, ended]
      responses:
        "200":
          description: List of auctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Auction"
    post:
      operationId: createAuction
      tags: [auctions]
      summary: Create a new auction (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuctionRequest"
      responses:
        "201":
          description: Auction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auctions/{id}:
    get:
      operationId: getAuction
      tags: [auctions]
      summary: Get auction by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Auction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAuction
      tags: [auctions]
      summary: Update auction (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAuctionRequest"
      responses:
        "200":
          description: Auction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
  /auctions/{id}/bids:
    get:
      operationId: getAuctionBids
      tags: [bids]
      summary: Get bids for an auction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of bids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bid"
    post:
      operationId: placeBid
      tags: [bids]
      summary: Place a bid
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBidRequest"
      responses:
        "201":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auctions/{id}/recommendation:
    get:
      operationId: getBidRecommendation
      tags: [auctions]
      summary: Get smart bid recommendation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bid recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BidRecommendation"
...
  /admin/users/{id}/credits:
    patch:
      operationId: assignCredits
      tags: [admin]
      summary: Assign credits to user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignCreditsRequest"
      responses:
        "200":
          description: Credits assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/generate-description:
    post:
      operationId: generateDescription
      tags: [admin]
      summary: AI-style auction description generator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDescriptionRequest"
      responses:
        "200":
          description: Generated description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateDescriptionResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, bidder]
        totalCredits:
          type: integer
        reservedCredits:
          type: integer
        availableCredits:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - totalCredits
        - reservedCredits
        - availableCredits
        - createdAt
    Auction:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        minBid:
          type: integer
        bidIncrement:
          type: integer
        currentBid:
          type: integer
        highestBidderId:
          type: integer
          nullable: true
        highestBidderName:
          type: string
          nullable: true
        status:
          type: string
          enum: [upcoming, active, ended]
        winnerId:
          type: integer
          nullable: true
        winnerName:
          type: string
          nullable: true
        totalBids:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - startTime
        - endTime
        - minBid
        - bidIncrement
        - currentBid
        - status
        - totalBids
        - createdAt
    AuctionDetail:
      allOf:
        - $ref: "#/components/schemas/Auction"
        - type: object
          properties:
            bids:
              type: array
              items:
                $ref: "#/components/schemas/Bid"
          required:
            - bids
    CreateAuctionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        minBid:
          type: integer
          minimum: 1
        bidIncrement:
          type: integer
          minimum: 1
      required:
        - title
        - description
        - startTime
        - endTime
        - minBid
        - bidIncrement
    UpdateAuctionRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
    PlaceBidRequest:
      type: object
      properties:
        amount:
          type: integer
          minimum: 1
      required:
        - amount
    Bid:
      type: object
      properties:
        id:
          type: integer
        auctionId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - auctionId
        - userId
        - userName
        - amount
        - createdAt
...
    BidRecommendation:
      type: object
      properties:
        suggestedBid:
          type: integer
        strategy:
          type: string
          enum: [safe, aggressive, moderate]
        reasoning:
          type: string
        confidence:
          type: integer
          minimum: 0
          maximum: 100
      required:
        - suggestedBid
        - strategy
        - reasoning
        - confidence
    AssignCreditsRequest:
      type: object
      properties:
        amount:
          type: integer
          minimum: 1
        action:
          type: string
          enum: [add, set]
          default: add
      required:
        - amount
...
    GenerateDescriptionRequest:
      type: object
      properties:
        title:
          type: string
        keywords:
          type: string
      required:
        - title
    GenerateDescriptionResponse:
      type: object
      properties:
        description:
          type: string
      required:
        - description