# lib/api-zod/src/generated/types/auction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidGenius API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuctionStatus } from "./auctionStatus";

export interface Auction {
  id: number;
  title: string;
  description: string;
  imageUrl?: string | null;
  startTime: Date;
  endTime: Date;
  minBid: number;
  bidIncrement: number;
  currentBid: number;
  highestBidderId?: number | null;
  highestBidderName?: string | null;
  status: AuctionStatus;
  winnerId?: number | null;
  winnerName?: string | null;
  totalBids: number;
  createdAt: Date;
}