# lib/api-zod/src/generated/types/createAuctionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidGenius API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAuctionRequest {
  title: string;
  description: string;
  imageUrl?: string | null;
  startTime: Date;
  endTime: Date;
  /** @minimum 1 */
  minBid: number;
  /** @minimum 1 */
  bidIncrement: number;
}