# lib/api-zod/src/generated/types/generateDescriptionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BidGenius API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateDescriptionResponse {
  description: string;
}